+++ packages/terra-form-select/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Removed
+  * Removed prop `placeholder` from all the variants of select.
+
+* Changed
+  * Updated color for placeholder.
+  * Updated default text in all variants for placeholder to be `- Select -`.
+
 ## 6.19.0 - (October 27, 2020)
 
 * Added